:recycle: refactor: streamline encoding process by removing redundant encodeTable function and updating usage in encodeBasicProperties
